-fixed warnings and errors reported by gcc 2.95 on alpha
git-svn-id: svn+ssh://svn.code.sf.net/p/opensips/svn/trunk@1402 689a6050-402a-0410-94f2-e92a70836424

+++ resolve.c
@@ -798,7 +798,7 @@ static inline void filter_and_sort_naptr( struct rdata** head_p,
 			goto skip;
 
 		DBG("DEBUG:filter_and_sort_naptr: found valid %.*s -> %s\n",
-			naptr->services_len,naptr->services, naptr->repl);
+			(int)naptr->services_len,naptr->services, naptr->repl);
 
 		/* this is a supported service -> add it according to order to the 
 		 * new head list */
@@ -825,7 +825,7 @@ static inline void filter_and_sort_naptr( struct rdata** head_p,
 		continue;
 skip:
 		DBG("DEBUG:filter_and_sort_naptr: skipping %.*s -> %s\n",
-			naptr->services_len, naptr->services, naptr->repl);
+			(int)naptr->services_len, naptr->services, naptr->repl);
 skip0:
 		l->next = out;
 		out = l;
